+++ src/Entity/Organization.php
@@ -7,16 +7,13 @@ namespace App\Entity;
 use App\Repository\OrganizationRepository;
 use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
+use ITKDev\EntityBundle\Audit\Attribute\Auditable;
 
 #[ORM\Entity(repositoryClass: OrganizationRepository::class)]
 #[ORM\Table(name: 'organization')]
-class Organization
+#[Auditable]
+class Organization extends AbstractEntity
 {
-    #[ORM\Id]
-    #[ORM\GeneratedValue]
-    #[ORM\Column]
-    private ?int $id = null;
-
     #[ORM\Column(length: 255)]
     private string $name;
 
@@ -37,14 +34,10 @@ class Organization
         array $emailDomains,
         string $defaultFramework,
     ) {
+        parent::__construct();
         $this->name = $name;
         $this->emailDomains = self::normaliseDomains($emailDomains);
         $this->defaultFramework = $defaultFramework;
-    }
-
-    public function getId(): ?int
-    {
-        return $this->id;
     }
 
     public function getName(): string